backend/requirements.txt
# Workspace version: 1.43.1
ipykernel==6.29.5
python-multipart==0.0.9
databutton==0.38.34
uvicorn[standard]==0.29.0
fastapi==0.111.0
openai
beautifulsoup4
requests
replicate
Pillow